[Dispatch] Removed derived-to-base conversion relicts
This functionality was used for the old DoubleDispatcher which operated only on polymorphic classes. This allowed to call functions on base class parameters, in case no match for derived classes was found, by performing an automatic derived-to-base conversion in the library. While this is a very useful feature, it has been inactive since the design of the more generalized DoubleDispatcher, thus this commit only removes dead code.

In case the feature should be re-integrated one day, consider also commit 11c2bf9633ec0d15f330a6319caecc9fd8db1bb1, before which the derived-to-base conversion was used in client code.